% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpipe.R
\name{check_and_unpipe}
\alias{check_and_unpipe}
\title{Check if call to \code{as_loop} used pipes, and if, unpipe it.}
\usage{
check_and_unpipe(sc, is_dot)
}
\description{
Check if call to \code{as_loop} used pipes, and if, unpipe it.
}
