% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_options.R
\name{create_output_fn}
\alias{create_output_fn}
\title{depending on output_context check and create final output function}
\usage{
create_output_fn(output)
}
\description{
depending on output_context check and create final output function
}
