% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_supported_fns}
\alias{get_supported_fns}
\title{Show a list of supported function names}
\usage{
get_supported_fns(fn)
}
\arguments{
\item{fn}{The name of a \code{{loopurrr}} function as string.}
}
\value{
A list of supported function names as named character vectors.
}
\description{
\code{get_supported_fns()} shows which functions are supported for a specific \code{{loopurrr}} function.
Currently, only works on \code{as_loop()}.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(width = 60)
get_supported_fns("as_loop")
#> $map
#>  [1] "map"     "map_at"  "map_chr" "map_dbl" "map_df" 
#>  [6] "map_dfc" "map_dfr" "map_if"  "map_int" "map_lgl"
#> [11] "map_raw"
#> 
#> $imap
#> [1] "imap"     "imap_chr" "imap_dbl" "imap_dfc" "imap_dfr"
#> [6] "imap_int" "imap_lgl" "imap_raw"
#> 
#> $map2
#> [1] "map2"     "map2_chr" "map2_dbl" "map2_df"  "map2_dfc"
#> [6] "map2_dfr" "map2_int" "map2_lgl" "map2_raw"
#> 
#> $pmap
#> [1] "pmap"     "pmap_chr" "pmap_dbl" "pmap_df"  "pmap_dfc"
#> [6] "pmap_dfr" "pmap_int" "pmap_lgl" "pmap_raw"
#> 
#> $lmap
#> [1] "lmap"    "lmap_at"
#> 
#> $modify
#> [1] "modify"    "modify_at" "modify_if" "modify2"  
#> [5] "imodify"  
#> 
#> $walk
#> [1] "iwalk" "pwalk" "walk"  "walk2"
#> 
#> $accumulate
#> [1] "accumulate"  "accumulate2"
#> 
#> $reduce
#> [1] "reduce"  "reduce2"
}\if{html}{\out{</div>}}
}

