% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopurrr-package.R
\docType{package}
\name{loopurrr-package}
\alias{loopurrr}
\alias{loopurrr-package}
\title{@description
Translate {purrr}'s iterator function to regular for-loops}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tim Tiefenbach \email{mailme@tim-tiefenbach.de} (\href{https://orcid.org/0000-0001-9443-2434}{ORCID})

}
\keyword{internal}
