% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_options.R
\name{get_output_opt}
\alias{get_output_opt}
\alias{set_output_opt}
\alias{default_context}
\title{Get and set output loopurrr's options}
\usage{
get_output_opt(default = NULL)

set_output_opt(x = list("rstudio", "clipboard", "console", NULL))

default_context()
}
\arguments{
\item{default}{If the specified option is not set in the options list, this value is returned.
This argument is for internal use only.}

\item{x}{Either \code{NULL} or one or several of \code{"rstudio"}, \code{"clipboard"}, \code{"console"}. If set to
more than one option, the output options will be run in order from left to right until
successful.}
}
\value{
For \code{get_output_opt()}, the current value set for option \code{"loopurrr.output"}, or default
(which defaults to NULL) if the option is unset.

For \code{default_context()}, either the current value set for option \code{"loopurrr.output"}.
In this case, if \code{"console"} is not among the options, it will be automatically included as last
option. Or, if option \code{"loopurrr.output"} is not specified \code{c("rstudio", "clipboard", "console")}.
}
\description{
\code{get_output_opt()} and \code{set_output_opt()} get and set loopurrr's output options. They are light
wrappers around \code{getOption("loopurrr.output")} and \code{options("loopurrr.output")}.

\code{default_context()} inspects if the \code{"loopurrr.output"} option is set. If the option is not
specified, it will default to \code{c("rstudio", "clipboard", "console")}. If \code{"console"} is not
among the output options it will be automatically included as last option.
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{set_output_opt(c("clipboard", "rstudio"))

get_ouptut_opt()
#> [1] "clipboard" "rstudio"

default_context()
#> [1] "clipboard" "rstudio" "console"
}\if{html}{\out{</div>}}
}

