% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopurrr-package.R
\docType{package}
\name{loopurrr-package}
\alias{loopurrr}
\alias{loopurrr-package}
\title{loopurrr: Tranlate purrr iterator functions to regular for loops}
\description{
Translate {purrr}'s iterator functions to regular for loops
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TimTeaFan/loopurrrr}
  \item Report bugs at \url{https://github.com/TimTeaFan/loopurrr/issues}
}

}
\author{
\strong{Maintainer}: Tim Tiefenbach \email{mailme@tim-tiefenbach.de} (\href{https://orcid.org/0000-0001-9443-2434}{ORCID})

}
\keyword{internal}
